+++ gemrb/tests/PVRZImporter/Test_PVRZImporter.cpp
@@ -51,10 +51,10 @@ TEST(PVRZImporterTest, GetBlockPixelMask1)
 	Region r { 1, 1, 2, 13 };
 	Region grid { 0, 0, 1, 4 };
 
-	EXPECT_EQ(26208, PVRZImporter::GetBlockPixelMask(r, grid, 0, 0));
-	EXPECT_EQ(26214, PVRZImporter::GetBlockPixelMask(r, grid, 0, 1));
-	EXPECT_EQ(26214, PVRZImporter::GetBlockPixelMask(r, grid, 0, 2));
-	EXPECT_EQ(102, PVRZImporter::GetBlockPixelMask(r, grid, 0, 3));
+	EXPECT_EQ(uint16_t(26208), PVRZImporter::GetBlockPixelMask(r, grid, 0, 0));
+	EXPECT_EQ(uint16_t(26214), PVRZImporter::GetBlockPixelMask(r, grid, 0, 1));
+	EXPECT_EQ(uint16_t(26214), PVRZImporter::GetBlockPixelMask(r, grid, 0, 2));
+	EXPECT_EQ(uint16_t(102), PVRZImporter::GetBlockPixelMask(r, grid, 0, 3));
 }
 
 /*
@@ -68,8 +68,8 @@ TEST(PVRZImporterTest, GetBlockPixelMask2)
 	Region r { 1, 1, 6, 2 };
 	Region grid { 0, 0, 2, 1 };
 
-	EXPECT_EQ(3808, PVRZImporter::GetBlockPixelMask(r, grid, 0, 0));
-	EXPECT_EQ(1904, PVRZImporter::GetBlockPixelMask(r, grid, 1, 0));
+	EXPECT_EQ(uint16_t(3808), PVRZImporter::GetBlockPixelMask(r, grid, 0, 0));
+	EXPECT_EQ(uint16_t(1904), PVRZImporter::GetBlockPixelMask(r, grid, 1, 0));
 }
 
 /*
@@ -83,7 +83,7 @@ TEST(PVRZImporterTest, GetBlockPixelMask3)
 	Region r { 1, 1, 2, 2 };
 	Region grid { 0, 0, 1, 1 };
 
-	EXPECT_EQ(1632, PVRZImporter::GetBlockPixelMask(r, grid, 0, 0));
+	EXPECT_EQ(uint16_t(1632), PVRZImporter::GetBlockPixelMask(r, grid, 0, 0));
 }
 
 /*
@@ -97,7 +97,7 @@ TEST(PVRZImporterTest, GetBlockPixelMask4)
 	Region r { 0, 0, 2, 2 };
 	Region grid { 0, 0, 1, 1 };
 
-	EXPECT_EQ(51, PVRZImporter::GetBlockPixelMask(r, grid, 0, 0));
+	EXPECT_EQ(uint16_t(51), PVRZImporter::GetBlockPixelMask(r, grid, 0, 0));
 }
 
 /*
@@ -111,6 +111,6 @@ TEST(PVRZImporterTest, GetBlockPixelMask5)
 	Region r { 2, 2, 2, 2 };
 	Region grid { 0, 0, 1, 1 };
 
-	EXPECT_EQ(52224, PVRZImporter::GetBlockPixelMask(r, grid, 0, 0));
+	EXPECT_EQ(uint16_t(52224), PVRZImporter::GetBlockPixelMask(r, grid, 0, 0));
 }
 }
